% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsAndStats.R
\name{extract_event_data}
\alias{extract_event_data}
\title{Extract a field from 'data' column of a data.frame of events}
\usage{
extract_event_data(events, field_name, column_name = field_name)
}
\arguments{
\item{events}{A data.frame, typically returned from the find method of a
mongo collection.}

\item{field_name}{A character vector containing of the name of the field to
extract.}
}
\description{
Attempts to extract a named member of the 'data' field as an atomic vector,
converting NULL to NA.  If any value in result has a length > 1, extracted
values are returned as a list.  Length of result will be equal to nrow(events).
}
