% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsAndStats.R
\name{identify_users}
\alias{identify_users}
\title{Looks up the most recent name associated with each user}
\usage{
identify_users(events_collection)
}
\arguments{
\item{events_collection}{Mongo collection objects pointing to the events
collection.}
}
\value{
A tibble of users with their associated name and the date/time when
  the name was entered.
}
\description{
Looks up the most recent name associated with each user
}
