% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{process_D2L_files}
\alias{process_D2L_files}
\alias{unzip_D2L_files}
\alias{check_D2L_files}
\alias{import_D2L_files}
\title{Process a zip file of assignment submissions from D2L}
\usage{
process_D2L_files(zip = file.choose(), overwrite = F)

unzip_D2L_files(zip = file.choose(), overwrite = F)

check_D2L_files(extracted_path)

import_D2L_files(extracted_path = dirname(file.choose()))
}
\arguments{
\item{zip}{Path to a D2L .zip file that contains tutorial submissions from
multiple users}

\item{overwrite}{When TRUE, will overwrite any existing folder to which the
.zip file has been extracted in the past.  When FALSE, the function fails
with an error if there is an existing folder with the assignment name.}

\item{extracted_path}{The path to a folder containing files that were
extracted by \code{unzip_D2L_files}.  The default value will invoke a file
chooser.  Select a file *within* the folder to be processed.  Although you
only select one file, all of the files in the same folder will imported. To
view the contents of just one file submitted by a student, see
\code{\link{tut_to_tibble}}}
}
\value{
\code{process_D2L_files} returns a tibble of all data for all
  students in the D2L .zip with, with a warning if any student's file was
  moved to the "bad_data" directory.

  \code{unzip_D2L_files()} returns the name of the new folder containing the
  unzipped files.  The folder name is the assignment name in D2L, which is
  extracted from the .zip file name.

  \code{check_D2L_files()} returns the \code{extracted_path} argument
  unaltered to facilitate piping the results of the command.

  \code{import_D2L_files()} returns a tibble containing all data from all
  files in \code{extracted_directory}.
}
\description{
In D2L, all users submissions for a single assignment can be downloaded in a
.zip file.  These functions process results from such a zip file into a
tibble.
}
\details{
\code{process_D2L_files} is a convenience wrapper that calls
\code{unzip_D2L_files()}, followed by \code{check_D2L_files()}, and
\code{import_D2L_files()}  on a .zip file downloaded from D2L.

\code{unzip_D2L_files()} will unzip a D2L .zip file into a separate directory
within the parent directory of the D2L file specified by \code{zip}.

\code{check_D2L_files()} will make sure all of the unzipped files can be
opened.  Any that can't be opened by R are placed into a "bad_files"
directory in the parent directory of the unzipped files with a warning.  In
this case the student may have to resubmit a file, or you might be able to
salvage the file manually and move it out of the "bad_files" folder.

\code{import_D2L_files()} will read in and process all of the unzipped
student-submitted files in a folder.
}
