% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsAndStats.R
\name{remove_user}
\alias{remove_user}
\title{Remove all events for a given user from the events table}
\usage{
remove_user(user, events_collection, commit = FALSE)
}
\arguments{
\item{user}{The user id to remove -- must match the value in the `user` key
in the collection}

\item{events_collection}{A mongo collection object pointing to the events
collection.}

\item{commit}{If FALSE, the function will only summarize events to be
removed; if TRUE, all events associated with `user` are removed.}
}
\value{
A tibble summary of events to be removed if commit is FALSE,
  otherwise TRUE if removal is successful.
}
\description{
Remove all events for a given user from the events table
}
