% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{tut_to_tibble}
\alias{tut_to_tibble}
\title{Convert a tutorial output list to a tibble}
\usage{
tut_to_tibble(x)
}
\value{
A tibble version of the data in \code{x}
}
\description{
Convert data lists from individual student tutorial submission files to a
tibble.
}
\details{
Individual student submission files can be read using the \code{\link{load}}
function, which will result in a variable \code{tutorial_data} being created
in the global environment.  \code{tutorial_data} will contain a sorta yucky
list.  Pass \code{tutorial_data} to \code{tut_to_tibble} to have the data put
into a yummy tibble format.
}
