% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialize.R
\name{tutorialize}
\alias{tutorialize}
\alias{tutorialize_setup}
\title{Create simple learnr tutorials from an "answer key" style RMarkdown document}
\usage{
tutorialize(package_dir = getwd(), choose = F)

tutorialize_setup(tutorial_name, package_dir = getwd())
}
\arguments{
\item{package_dir}{The root directory of the package that contains answer key
files to be tutorialized}
}
\description{
Set up the structure necessary to create tutorialized learnr tutorials and
create them
}
\details{
To create a new tutorial using tutorialize:

\itemize{

  \item create a package project

  \item from within the package project in R Studio, run
\code{usethis::use_tutorial} to create a straw learnr tutorial.  The tutorial
file will be created in an /inst/tutorials directory in your package project.

  \item run tutorialize_setup using the directory name of the tutorial.  This
will create an "answer_key" directory within the tutorial's directory and
move the straw learnr file to the "answer_key" directory.

  \item edit the tutorial file in the "answer_key" directory using the
  tips below.

  \item When you are ready to tutorialize your answer_keys, run the
\code{tutorialize} which will recursively search for any answer_keys and
process them, creating a learnr tutorial file of the same name up one
directory from the answer key file.
}

TIPS

\code{tutorialize} will recognize and process any chuck set with
"exercise=tutorialize" as an option.  If this option is set:

\itemize{

\item Any code in the chunk will be moved to the learn r "solution" block

\item Any text prefaced by `#+` at the beginning of the line will not be
moved to the setup chuck, but will be stripped of `#+` and will remain as
"starter" code for the exercise

\item A learn "code-check" block will be added, containing a call to
\code{\link{grade_code()}}

}

A custom learnr-style "setup" chunk can be used before any tutorialze chunk.
if "exercise-setup=foo" is included as an option for a tutorialize chunk,
then the setup chunk for code chunk foo and any assignments made in the
solution chunk for foo will be added to the current chunk's setup chunk. This
allows for chaining of tutorialized chunks.

Manually created learnr chunks (those with "exercise=TRUE" option) can be
intermingled with tutorialize chunks.  The learnr chunks and will be passed
through tutorialize unaltered.
}
