% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsAndStats.R
\name{fetch_events}
\alias{fetch_events}
\alias{event_stats}
\title{Get events and behavioral statistics for each user and tutorial response.}
\usage{
fetch_events(events_collection)

event_stats(events)
}
\arguments{
\item{events_collection}{Mongo collection object for events collection}

\item{events}{A tibble derived from \code{fetch_events()}}
}
\value{
A tibble of statistics for each survey question.
}
\description{
Get events and behavioral statistics for each user and tutorial response.
}
